chore(schema): migrate the data_content column on the file system
This column was previously used to store JSON content, but the size of this data
can be big, so the solution is to move this data on the file system instead.

The migration will create the storage directory, move the content in dedicated
files, write the filename path in the data_path column for each entries and
remove the data_content column.